Changed weight from int to float

+++ Assets/Scripts/ItemData.cs
@@ -12,7 +12,7 @@ public class ItemData : ScriptableObject
     public bool isStatTrak;
     public string type;
     public string rarity;
-    public int weight;
+    public float weight;
     public int demandScore;
 
     public float lastActivityTime;
